Remove Vim configuration
I've been using NeoVim for more than a year now. So far so good, and I'm
pretty much happy about that. Since my Vim configuration wasn't
maintained for quite a while, I don't want to even pretend that it's
working. So I'm removing this for greater good.

P.S: It should not be a big deal to adapt my NeoVim configuration to run
in Vim.