yay: add configuration to vcs
Yay is AUR helper written in Go. Yes, this is a silly language, but it
seems well maintained and hence could be used to install things from
AUR. Yay comes with a silly default when PKGBUILD diffs aren't shown.
This configuration enables showing diffs when you install or update
packages.